kafka-redis-consumer/requirements.txt
kafka-python==2.0.2
redis==5.0.1
python-json-logger==2.0.7 